Code for turning a motor on and off
This code is an infinite loop that sends a signal over a GPIO pin to a relay that opens and closes a separate circuit that allows for a hobby motor to be turned on and off. The code is in no way polished and is instead simply a progress for where I am at currently.